backend/requirements.txt
# API Framework
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
httpx>=0.27.0
aiofiles>=23.2.1
pydantic>=2.7.1

# Database & Config
motor>=3.4.0
pymongo[srv]>=4.7.2
python-dotenv>=1.0.1
redis>=5.0.4

# Computer Vision Engine
opencv-python-headless>=4.9.0
ultralytics>=8.0.0
Pillow>=10.3.0

# Demand Forecasting & Analytics
scikit-learn>=1.4.2
numpy>=1.26.4
pandas>=2.2.2
statsmodels>=0.14.2

# NOTE: Prophet is supported by our pipeline architecture as requested by 
# the DAIICT Problem Statement, but is omitted from local hackathon 
# builds to prevent C++/PyStan environment corruption on Windows.
# statsmodels Holt-Winters is the default edge-backend.
# prophet>=1.1.5
